# packages/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: CineWeave API
  version: 0.1.0
  description: API contract seed for the Phase 1 CineWeave control plane.
servers:
  - url: http://localhost:8080
paths:
...
  /api/projects/{projectId}/regenerate:
    post:
      operationId: regenerateCreativeObject
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegenerateRequest"
      responses:
        "202":
          description: Partial regeneration workflow queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegenerateResponseEnvelope"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
...
  /api/projects/{projectId}/shot-asset-requirements/{requirementId}:
    patch:
      operationId: updateShotAssetRequirement
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/RequirementId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateShotAssetRequirementRequest"
      responses:
        "200":
          description: Updated shot asset requirement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShotAssetRequirementEnvelope"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
...
  /api/projects/{projectId}/storyboard-shots/{shotId}:
    patch:
      operationId: updateStoryboardShot
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/ShotId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStoryboardShotRequest"
      responses:
        "200":
          description: Updated storyboard shot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoryboardShotEnvelope"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    serviceTokenAuth:
      type: http
      scheme: bearer
      bearerFormat: service-token
  parameters:
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    RequirementId:
      name: requirementId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ShotId:
      name: shotId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
    Envelope:
      type: object
      required: [requestId]
      properties:
        requestId:
          type: string
        data: {}
        meta: {}
...
    ErrorEnvelope:
      type: object
      required: [requestId, error]
      properties:
        requestId:
          type: string
        error:
          $ref: "#/components/schemas/ErrorBody"
    ErrorBody:
      type: object
      required: [code, message, retryable]
      properties:
        code:
          type: string
          description: Includes PROVIDER_RATE_LIMITED, PROVIDER_CONCURRENCY_LIMITED, PROVIDER_DAILY_QUOTA_EXCEEDED, PROVIDER_MONTHLY_BUDGET_EXCEEDED, and PROVIDER_CIRCUIT_OPEN for Provider Gateway guard blocks.
        message:
          type: string
        details: {}
        retryable:
          type: boolean
...
    ShotAssetRequirement:
      type: object
      required: [id, organizationId, projectId, storyboardShotId, assetId, requirementType, status, reviewStatus, manualOverride, staleState, metadata, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        workflowRunId:
          type: string
          format: uuid
        storyboardShotId:
          type: string
          format: uuid
        assetId:
          type: string
          format: uuid
        requirementType:
          type: string
        roleInShot:
          type: string
        costume:
          type: string
        pose:
          type: string
        expression:
          type: string
        action:
          type: string
        cameraRelation:
          type: string
        sceneState:
          type: string
        propState:
          type: string
        prompt:
          type: string
        derivedArtifactId:
          type: string
          format: uuid
        derivedMediaFileId:
          type: string
          format: uuid
        derivedStorageKey:
          type: string
        status:
          type: string
        reviewStatus:
          $ref: "#/components/schemas/ReviewStatus"
        manualOverride:
          type: boolean
        staleState:
          $ref: "#/components/schemas/StaleState"
        editedBy:
          type: string
          format: uuid
        editedAt:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    StoryboardShot:
      type: object
      required: [id, workflowRunId, shotIndex, shotNo, status, reviewStatus, manualOverride, staleState]
      properties:
        id:
          type: string
          format: uuid
        workflowRunId:
          type: string
          format: uuid
        scriptSceneId:
          type: string
          format: uuid
        sourceScene:
          $ref: "#/components/schemas/ShotSourceScene"
        shotIndex:
          type: integer
        shotNo:
          type: integer
        durationSeconds:
          type: number
          format: double
        visual:
          type: string
        camera:
          type: string
        motion:
          type: string
        mood:
          type: string
        imagePrompt:
          type: string
        videoPrompt:
          type: string
        imageArtifactId:
          type: string
          format: uuid
        imageMediaFileId:
          type: string
          format: uuid
        imageStorageKey:
          type: string
        imagePreviewUrl:
          type: string
          format: uri
        videoArtifactId:
          type: string
          format: uuid
        videoMediaFileId:
          type: string
          format: uuid
        videoStorageKey:
          type: string
        videoPreviewUrl:
          type: string
          format: uri
        providerAsyncTaskId:
          type: string
          format: uuid
        externalTaskId:
          type: string
        status:
          type: string
        reviewStatus:
          $ref: "#/components/schemas/ReviewStatus"
        manualOverride:
          type: boolean
        staleState:
          $ref: "#/components/schemas/StaleState"
        editedBy:
          type: string
          format: uuid
        editedAt:
          type: string
          format: date-time
    ShotSourceScene:
      type: object
      required: [id, sceneNo, title, characters]
      properties:
        id:
          type: string
          format: uuid
        sceneNo:
          type: integer
        title:
          type: string
        location:
          type: string
        characters:
          type: array
          items:
            type: string
...
    ReviewStatus:
      type: string
      enum: [pending, approved, rejected, needs_edit]
    StaleState:
      type: string
      enum: [fresh, upstream_changed, needs_regeneration]
...
    RegenerateRequest:
      type: object
      required: [targetType]
      properties:
        targetType:
          type: string
          enum: [canonical_asset_image, derived_asset_image, shot_image, shot_video, final_video, script_scene, scene_storyboard]
        targetId:
          type: string
          format: uuid
          description: Required for all targets except final_video. For final_video it optionally selects the source workflow run.
        options:
          type: object
          additionalProperties: true
    RegenerateResponse:
      type: object
      required: [targetType, targetId, workflowRunId, status, workflowType]
      properties:
        targetType:
          type: string
        targetId:
          type: string
          format: uuid
        workflowRunId:
          type: string
          format: uuid
        status:
          type: string
        workflowType:
          type: string
...
    RegenerateResponseEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/RegenerateResponse"
...
    ShotAssetRequirementEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ShotAssetRequirement"
...
    StoryboardShotEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/StoryboardShot"
...
    UpdateShotAssetRequirementRequest:
      type: object
      properties:
        costume:
          type: string
        pose:
          type: string
        expression:
          type: string
        action:
          type: string
        cameraRelation:
          type: string
        sceneState:
          type: string
        propState:
          type: string
        prompt:
          type: string
    UpdateStoryboardShotRequest:
      type: object
      properties:
        visual:
          type: string
        camera:
          type: string
        motion:
          type: string
        mood:
          type: string
        durationSeconds:
          type: number
          format: double
          minimum: 0
        imagePrompt:
          type: string
        videoPrompt:
          type: string
...